refactoring continues

+++ src/main/java/com/org/pos/repository/VentasRepository.java
@@ -658,7 +658,7 @@ public class VentasRepository {
           productoParaCarrrito.setPrecioVenta(rs.getDouble("precioUnitarioV"));
           productoParaCarrrito.setPresentacion(rs.getString("presentacion"));
           productoParaCarrrito.setUnidadMedida(rs.getString("uMedida"));
-          productoParaCarrrito.setUnidadesEnCaja(rs.getInt("unidadesEnCaja"));
+          productoParaCarrrito.setUnidadesEnCaja(rs.getDouble("unidadesEnCaja"));
           
           tipoDeProductoAgregado=rs.getInt("TipoProducto");
           
@@ -726,8 +726,8 @@ public class VentasRepository {
 
          if(contador>0){
              
-            int cantidadAvender=0;//Integer.parseInt(cantidadPizza.getText());
-            int cantidadEnAlmacen=productoParaCarrrito.getUnidadesEnCaja();
+            Double cantidadAvender=0.0;//Integer.parseInt(cantidadPizza.getText());
+            Double cantidadEnAlmacen=productoParaCarrrito.getUnidadesEnCaja();
             ///recorrer y sumar los productos con el mismo codigo o id 
             ///para obtener cuantos productos en total de cantidad se han agregado
